*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@1272 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/vertmodel.h
@@ -270,7 +270,7 @@ struct vertmodel : model
                 vector<animinfo> &ais = anims[anim&ANIM_INDEX];
                 if(ais.length())
                 {
-                    animinfo &ai = ais[varseed%ais.length()];
+                    animinfo &ai = ais[uint(varseed)%ais.length()];
                     as.frame = ai.frame;
                     as.range = ai.range;
                     if(ai.speed>0) as.speed = 1000.0f/ai.speed;
